Tests: Moved ContextualToolbar integration tests to a separate test file.
